move types to shared directory
