+++ src/shared/types/coords.interface.ts
