+++ src/shared/types/housing-features.type.ts
