+++ src/shared/types/mock-server-data.type.ts
