+++ src/shared/types/rental-offer.interface.ts
