+++ src/shared/types/type-housing.type.ts
