+++ src/shared/types/user.interface.ts
